new class (system manager)

+++ Ocean/player/player.cpp
@@ -46,7 +46,7 @@ Player::~Player()
 }
 
 //Public slots
-void Player::CallSetMod(const unsigned short int &mod)
+void Player::CallSetMod(const int &mod)
 {
     if(Player::SetModOfPlayer(mod))
         qDebug() << "set mod: " << mod;
@@ -66,7 +66,7 @@ void Player::CallSetPlayList(QMediaPlaylist *playlist)
     return;
 }
 
-void Player::CallSetVolume(const unsigned short int &volume)
+void Player::CallSetVolume(const int &volume)
 {
     if(Player::SetVolume(volume))
         qDebug() << "set volume" << volume;
@@ -103,7 +103,7 @@ void Player::SetPositionOfTrack(const qint64 position)
     return;
 }
 
-bool Player::SetModOfPlayer(const unsigned short int &mod)
+bool Player::SetModOfPlayer(const int &mod)
 {
     switch(mod)
     {
@@ -146,7 +146,7 @@ bool Player::SetPlayList(QMediaPlaylist *playlist)
         return false;
 }
 
-bool Player::SetVolume(const unsigned short int &volume)
+bool Player::SetVolume(const int &volume)
 {
     if(volume > 0)
     {
